take lists out of events
