+++ electricitymap/contrib/libs/tests/test_events.py
@@ -8,17 +8,12 @@ from mock import patch
 from electricitymap.contrib.config import ZoneKey
 from electricitymap.contrib.libs.models.events import (
     Exchange,
-    ExchangeList,
     Price,
-    PriceList,
     ProductionBreakdown,
-    ProductionBreakdownList,
     ProductionMix,
     StorageMix,
     TotalConsumption,
-    TotalConsumptionList,
     TotalProduction,
-    TotalProductionList,
 )
 
 
@@ -42,22 +37,6 @@ class TestExchange(unittest.TestCase):
             source="trust.me",
         )
         assert exchange.value == -1
-
-    def test_exchange_list(self):
-        exchange_list = ExchangeList(logging.Logger("test"))
-        exchange_list.append(
-            zoneKey=ZoneKey("AT->DE"),
-            datetime=datetime(2023, 1, 1, tzinfo=timezone.utc),
-            value=1,
-            source="trust.me",
-        )
-        exchange_list.append(
-            zoneKey=ZoneKey("AT->DE"),
-            datetime=datetime(2023, 1, 2, tzinfo=timezone.utc),
-            value=1,
-            source="trust.me",
-        )
-        assert len(exchange_list.events) == 2
 
     def test_raises_if_invalid_exchange(self):
         with self.assertRaises(ValueError):
@@ -101,17 +80,6 @@ class TestExchange(unittest.TestCase):
                 source="trust.me",
             )
 
-    def test_append_to_list_logs_error(self):
-        exchange_list = ExchangeList(logging.Logger("test"))
-        with patch.object(exchange_list.logger, "error") as mock_error:
-            exchange_list.append(
-                zoneKey=ZoneKey("AT"),
-                datetime=datetime(2023, 1, 1, tzinfo=timezone.utc),
-                value=1,
-                source="trust.me",
-            )
-            mock_error.assert_called_once()
-
 
 class TestConsumption(unittest.TestCase):
     def test_create_consumption(self):
@@ -148,33 +116,6 @@ class TestConsumption(unittest.TestCase):
                 consumption=-1,
                 source="trust.me",
             )
-
-    def test_consumption_list(self):
-        consumption_list = TotalConsumptionList(logging.Logger("test"))
-        consumption_list.append(
-            zoneKey=ZoneKey("AT"),
-            datetime=datetime(2023, 1, 1, tzinfo=timezone.utc),
-            consumption=1,
-            source="trust.me",
-        )
-        consumption_list.append(
-            zoneKey=ZoneKey("AT"),
-            datetime=datetime(2023, 1, 2, tzinfo=timezone.utc),
-            consumption=1,
-            source="trust.me",
-        )
-        assert len(consumption_list.events) == 2
-
-    def test_append_to_list_logs_error(self):
-        consumption_list = TotalConsumptionList(logging.Logger("test"))
-        with patch.object(consumption_list.logger, "error") as mock_error:
-            consumption_list.append(
-                zoneKey=ZoneKey("AT"),
-                datetime=datetime(2023, 1, 1, tzinfo=timezone.utc),
-                consumption=-1,
-                source="trust.me",
-            )
-            mock_error.assert_called_once()
 
 
 class TestPrice(unittest.TestCase):
@@ -217,29 +158,6 @@ class TestPrice(unittest.TestCase):
                 source="trust.me",
                 currency="EURO",
             )
-
-    def test_price_list(self):
-        price_list = PriceList(logging.Logger("test"))
-        price_list.append(
-            zoneKey=ZoneKey("AT"),
-            datetime=datetime(2023, 1, 1, tzinfo=timezone.utc),
-            price=1,
-            source="trust.me",
-            currency="EUR",
-        )
-        assert len(price_list.events) == 1
-
-    def test_append_to_list_logs_error(self):
-        price_list = PriceList(logging.Logger("test"))
-        with patch.object(price_list.logger, "error") as mock_error:
-            price_list.append(
-                zoneKey=ZoneKey("AT"),
-                datetime=datetime(2023, 1, 1, tzinfo=timezone.utc),
-                price=1,
-                source="trust.me",
-                currency="EURO",
-            )
-            mock_error.assert_called_once()
 
 
 class TestProductionBreakdown(unittest.TestCase):
@@ -323,35 +241,6 @@ class TestProductionBreakdown(unittest.TestCase):
             assert breakdown.production.hydro == None
             assert breakdown.production.wind == 10
 
-    def test_production_list(self):
-        production_list = ProductionBreakdownList(logging.Logger("test"))
-        production_list.append(
-            zoneKey=ZoneKey("AT"),
-            datetime=datetime(2023, 1, 1, tzinfo=timezone.utc),
-            production=ProductionMix(wind=10),
-            source="trust.me",
-        )
-        assert len(production_list.events) == 1
-
-    def test_production_list_logs_error(self):
-        production_list = ProductionBreakdownList(logging.Logger("test"))
-        with patch.object(production_list.logger, "error") as mock_error:
-            production_list.append(
-                zoneKey=ZoneKey("AT"),
-                datetime=datetime(2023, 1, 1),
-                production=ProductionMix(wind=10),
-                source="trust.me",
-            )
-            mock_error.assert_called_once()
-        with patch.object(production_list.logger, "warning") as mock_warning:
-            production_list.append(
-                zoneKey=ZoneKey("AT"),
-                datetime=datetime(2023, 1, 1, tzinfo=timezone.utc),
-                production=ProductionMix(wind=-10),
-                source="trust.me",
-            )
-            mock_warning.assert_called_once()
-
     @freezegun.freeze_time("2023-01-01")
     def test_forecasted_points(self):
         mix = ProductionMix(wind=10)
@@ -392,13 +281,3 @@ class TestTotalProduction(unittest.TestCase):
         assert generation.datetime == datetime(2023, 1, 1, tzinfo=timezone.utc)
         assert generation.source == "trust.me"
         assert generation.value == 1
-
-    def test_generation_list(self):
-        generation_list = TotalProductionList(logging.Logger("test"))
-        generation_list.append(
-            zoneKey=ZoneKey("AT"),
-            datetime=datetime(2023, 1, 1, tzinfo=timezone.utc),
-            value=1,
-            source="trust.me",
-        )
-        assert len(generation_list.events) == 1
